test: Split models.export unittests
