+++ test/test_models.py
@@ -19,49 +19,9 @@ from doctr import models
 
 
 @pytest.fixture(scope="module")
-def mock_model():
-    _layers = [
-        layers.Conv2D(8, 3, activation='relu', padding='same', input_shape=(224, 224, 3)),
-        layers.GlobalAveragePooling2D(),
-        layers.Flatten(),
-        layers.Dense(10),
-    ]
-    return Sequential(_layers)
-
-
-@pytest.fixture(scope="module")
 def mock_vocab():
     return ('3K}7eé;5àÎYho]QwV6qU~W"XnbBvcADfËmy.9ÔpÛ*{CôïE%M4#ÈR:g@T$x?0î£|za1ù8,OG€P-kçHëÀÂ2É/ûIJ\'j'
             '(LNÙFut[)èZs+&°Sd=Ï!<â_Ç>rêi`l')
-
-
-@pytest.fixture(scope="module")
-def test_convert_to_tflite(mock_model):
-    serialized_model = models.export.convert_to_tflite(mock_model)
-    assert isinstance(serialized_model, bytes)
-    return serialized_model
-
-
-@pytest.fixture(scope="module")
-def test_convert_to_fp16(mock_model):
-    serialized_model = models.export.convert_to_fp16(mock_model)
-    assert isinstance(serialized_model, bytes)
-    return serialized_model
-
-
-@pytest.fixture(scope="module")
-def test_quantize_model(mock_model):
-    serialized_model = models.export.quantize_model(mock_model, (224, 224, 3))
-    assert isinstance(serialized_model, bytes)
-    return serialized_model
-
-
-def test_export_sizes(test_convert_to_tflite, test_convert_to_fp16, test_quantize_model):
-    assert sys.getsizeof(test_convert_to_tflite) > sys.getsizeof(test_convert_to_fp16)
-    if tf.__version__ < "2.4.0":
-        assert sys.getsizeof(test_convert_to_fp16) >= sys.getsizeof(test_quantize_model)
-    else:
-        assert sys.getsizeof(test_convert_to_fp16) > sys.getsizeof(test_quantize_model)
 
 
 def test_detpreprocessor(mock_pdf):  # noqa: F811
