+++ test/test_models_export.py
@@ -1,0 +1,46 @@
+import pytest
+import sys
+import tensorflow as tf
+from tensorflow.keras import layers, Sequential
+
+from doctr.models import export
+
+
+@pytest.fixture(scope="module")
+def mock_model():
+    _layers = [
+        layers.Conv2D(8, 3, activation='relu', padding='same', input_shape=(224, 224, 3)),
+        layers.GlobalAveragePooling2D(),
+        layers.Flatten(),
+        layers.Dense(10),
+    ]
+    return Sequential(_layers)
+
+
+@pytest.fixture(scope="module")
+def test_convert_to_tflite(mock_model):
+    serialized_model = export.convert_to_tflite(mock_model)
+    assert isinstance(serialized_model, bytes)
+    return serialized_model
+
+
+@pytest.fixture(scope="module")
+def test_convert_to_fp16(mock_model):
+    serialized_model = export.convert_to_fp16(mock_model)
+    assert isinstance(serialized_model, bytes)
+    return serialized_model
+
+
+@pytest.fixture(scope="module")
+def test_quantize_model(mock_model):
+    serialized_model = export.quantize_model(mock_model, (224, 224, 3))
+    assert isinstance(serialized_model, bytes)
+    return serialized_model
+
+
+def test_export_sizes(test_convert_to_tflite, test_convert_to_fp16, test_quantize_model):
+    assert sys.getsizeof(test_convert_to_tflite) > sys.getsizeof(test_convert_to_fp16)
+    if tf.__version__ < "2.4.0":
+        assert sys.getsizeof(test_convert_to_fp16) >= sys.getsizeof(test_quantize_model)
+    else:
+        assert sys.getsizeof(test_convert_to_fp16) > sys.getsizeof(test_quantize_model)
